requirements.txt
Django == 1.9
gunicorn==19.3.0
whitenoise==2.0.4
psycopg2==2.6.1
reportlab==3.2.0
taggit==0.17.6
Markdown==2.6.5
django-haystack==2.4.1
pysolr==3.3.3


